apps/backend/requirements/base.txt
# Core Framework
fastapi==0.115.14
fastapi-cli==0.0.7
uvicorn==0.35.0
starlette==0.46.2

# Database
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.2

# Data Validation & Serialization
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydantic-extra-types==2.10.5
email-validator==2.2.0
annotated-types==0.7.0

# Authentication & Security
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
python-multipart==0.0.20
cryptography==45.0.4
rsa==4.9.1
ecdsa==0.19.1
pyasn1==0.6.1

# Environment Variables
python-dotenv==1.1.1

# HTTP & Web
httpx==0.28.1
httpcore==1.0.9
httptools==0.6.4
h11==0.16.0
anyio==4.9.0
sniffio==1.3.1
idna==3.10
certifi==2025.6.15

# Performance
orjson==3.10.18
ujson==5.10.0
uvloop==0.21.0
websockets==15.0.1
watchfiles==1.1.0

# Template Engine
Jinja2==3.1.6
MarkupSafe==3.0.2
Mako==1.3.10

# CLI & Utilities
typer==0.16.0
click==8.2.1
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
Pygments==2.19.2
markdown-it-py==3.0.0
mdurl==0.1.2

# Configuration
PyYAML==6.0.2
itsdangerous==2.2.0

# Type checking
typing-extensions==4.14.0
typing-inspection==0.4.1

# Network utilities
dnspython==2.7.0

# C Extensions
cffi==1.17.1
pycparser==2.22

# Compatibility
six==1.17.0
